chatbot/telegram_ui/requirements.txt
asyncio
python-dotenv
aiogram
redis
requests